# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_monitoring-system_backend_pkg_response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: dev@company.com
    name: Internal Dev Team
  description: API untuk sistem monitoring audit internal perusahaan
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Monitoring Audit API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      summary: User Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate user session
      parameters:
      - description: Login Log ID
        in: query
        name: login_log_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get current user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - Auth
...
  /roles:
    get:
      consumes:
      - application/json
      description: Get paginated list of roles
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Role data
        in: body
        name: body
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - Roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: Get a role by its ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update an existing role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Role data
        in: body
        name: body
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Roles
  /roles/{id}/permissions:
    get:
      consumes:
      - application/json
      description: Get all permissions assigned to a role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Role Permissions
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Set or update permissions for a role in bulk
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Bulk Set Request
        in: body
        name: body
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_monitoring-system_backend_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: Bulk Set Role Permissions
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"